+++ src/backend/server-module.loader.ts
@@ -12,6 +12,9 @@ import {PagesDataserviceModule} from '@dps/mycms-server-commons/dist/pdoc-backen
 import {StaticPagesDataService} from '@dps/mycms-commons/dist/pdoc-commons/services/staticpages-data.service';
 import {PDocDataServiceModule} from '@dps/mycms-server-commons/dist/pdoc-backend-commons/modules/pdoc-dataservice.module';
 import {PDocServerModule} from '@dps/mycms-server-commons/dist/pdoc-backend-commons/modules/pdoc-server.module';
+import {MarkdownService} from '@dps/mycms-commons/dist/markdown-commons/markdown.service';
+import {DefaultOptions} from '@dps/mycms-commons/dist/markdown-commons/options';
+import {MarkdownDefaultExtensions} from '@dps/mycms-commons/dist/markdown-commons/extensions/markdown.extensions';
 
 export interface ServerConfig extends CommonServerConfigType<BackendConfigType, FirewallConfig> {
 }
@@ -39,10 +42,11 @@ export class ServerModuleLoader {
     }
 
     public static loadModulePages(app, serverConfig: ServerConfig, cache: DataCacheModule) {
+        const markdownService = new MarkdownService(DefaultOptions.getDefault(), MarkdownDefaultExtensions);
         const pagesDataServiceDE: StaticPagesDataService = PagesDataserviceModule.getDataService('pdocSolrDE',
-            serverConfig.backendConfig, 'de');
+            serverConfig.backendConfig, 'de', markdownService);
         const pagesDataServiceEN: StaticPagesDataService = PagesDataserviceModule.getDataService('pdocSolrEN',
-            serverConfig.backendConfig, 'en');
+            serverConfig.backendConfig, 'en', markdownService);
 
         PagesServerModule.configureRoutes(app, serverConfig.apiDataPrefix, pagesDataServiceDE, 'de', serverConfig.backendConfig.profile);
         PagesServerModule.configureRoutes(app, serverConfig.apiDataPrefix, pagesDataServiceEN, 'en', serverConfig.backendConfig.profile);