+++ docs/CHANGELOG.md
@@ -8,10 +8,12 @@
 - added optional password-encryption of static-viewer-version
 - backend/frontend: add page-management
 - improve imageInlineLoader
+- backend/frontend: use new fully featured markdown-support with extensions
 
 ### new features
 - added optional password-encryption of static-viewer-version
 - backend/frontend: add page-management
+- backend/frontend: use new fully featured markdown-support with extensions
 
 ### improvements
 - use commons 6.0.0